# Part page loading improvements