--- a/InvenTree/InvenTree/static/fullcalendar/locales/en-gb.js
+++ b/InvenTree/InvenTree/static/fullcalendar/locales/en-gb.js
@@ -7,6 +7,16 @@ FullCalendar.globalLocales.push(function () {
       dow: 1, // Monday is the first day of the week.
       doy: 4, // The week that contains Jan 4th is the first week of the year.
     },
+    buttonHints: {
+      prev: 'Previous $0',
+      next: 'Next $0',
+      today: 'This $0',
+    },
+    viewHint: '$0 view',
+    navLinkHint: 'Go to $0',
+    moreLinkHint(eventCnt) {
+      return `Show ${eventCnt} more event${eventCnt === 1 ? '' : 's'}`
+    },
   };
 
   return enGb;