--- a/InvenTree/InvenTree/static/fullcalendar/locales/es.js
+++ b/InvenTree/InvenTree/static/fullcalendar/locales/es.js
@@ -16,10 +16,29 @@ FullCalendar.globalLocales.push(function () {
       day: 'Día',
       list: 'Agenda',
     },
+    buttonHints: {
+      prev: '$0 antes',
+      next: '$0 siguiente',
+      today(buttonText) {
+        return (buttonText === 'Día') ? 'Hoy' :
+          ((buttonText === 'Semana') ? 'Esta' : 'Este') + ' ' + buttonText.toLocaleLowerCase()
+      },
+    },
+    viewHint(buttonText) {
+      return 'Vista ' + (buttonText === 'Semana' ? 'de la' : 'del') + ' ' + buttonText.toLocaleLowerCase()
+    },
     weekText: 'Sm',
+    weekTextLong: 'Semana',
     allDayText: 'Todo el día',
     moreLinkText: 'más',
+    moreLinkHint(eventCnt) {
+      return `Mostrar ${eventCnt} eventos más`
+    },
     noEventsText: 'No hay eventos para mostrar',
+    navLinkHint: 'Ir al $0',
+    closeHint: 'Cerrar',
+    timeHint: 'La hora',
+    eventHint: 'Evento',
   };
 
   return es;