--- a/InvenTree/InvenTree/static/fullcalendar/locales/hu.js
+++ b/InvenTree/InvenTree/static/fullcalendar/locales/hu.js
@@ -14,7 +14,7 @@ FullCalendar.globalLocales.push(function () {
       month: 'Hónap',
       week: 'Hét',
       day: 'Nap',
-      list: 'Napló',
+      list: 'Lista',
     },
     weekText: 'Hét',
     allDayText: 'Egész nap',