--- a/InvenTree/InvenTree/static/fullcalendar/locales/nb.js
+++ b/InvenTree/InvenTree/static/fullcalendar/locales/nb.js
@@ -17,9 +17,20 @@ FullCalendar.globalLocales.push(function () {
       list: 'Agenda',
     },
     weekText: 'Uke',
+    weekTextLong: 'Uke',
     allDayText: 'Hele dagen',
     moreLinkText: 'til',
     noEventsText: 'Ingen hendelser å vise',
+    buttonHints: {
+      prev: 'Forrige $0',
+      next: 'Neste $0',
+      today: 'Nåværende $0',
+    },
+    viewHint: '$0 visning',
+    navLinkHint: 'Gå til $0',
+    moreLinkHint(eventCnt) {
+      return `Vis ${eventCnt} flere hendelse${eventCnt === 1 ? '' : 'r'}`
+    },
   };
 
   return nb;