--- a/InvenTree/InvenTree/static/fullcalendar/locales/sm.js
+++ b/InvenTree/InvenTree/static/fullcalendar/locales/sm.js
@@ -1,0 +1,23 @@
+FullCalendar.globalLocales.push(function () {
+  'use strict';
+
+  var sm = {
+    code: 'sm',
+    buttonText: {
+      prev: 'Talu ai',
+      next: 'Mulimuli atu',
+      today: 'Aso nei',
+      month: 'Masina',
+      week: 'Vaiaso',
+      day: 'Aso',
+      list: 'Faasologa',
+    },
+    weekText: 'Vaiaso',
+    allDayText: 'Aso atoa',
+    moreLinkText: 'sili atu',
+    noEventsText: 'Leai ni mea na tutupu',
+  };
+
+  return sm;
+
+}());