--- a/InvenTree/InvenTree/static/fullcalendar/locales/sv.js
+++ b/InvenTree/InvenTree/static/fullcalendar/locales/sv.js
@@ -16,10 +16,30 @@ FullCalendar.globalLocales.push(function () {
       day: 'Dag',
       list: 'Program',
     },
+    buttonHints: {
+      prev(buttonText) {
+        return `Föregående ${buttonText.toLocaleLowerCase()}`
+      },
+      next(buttonText) {
+        return `Nästa ${buttonText.toLocaleLowerCase()}`
+      },
+      today(buttonText) {
+        return (buttonText === 'Program' ? 'Detta' : 'Denna') + ' ' + buttonText.toLocaleLowerCase()
+      },
+    },
+    viewHint: '$0 vy',
+    navLinkHint: 'Gå till $0',
+    moreLinkHint(eventCnt) {
+      return `Visa ytterligare ${eventCnt} händelse${eventCnt === 1 ? '' : 'r'}`
+    },
     weekText: 'v.',
+    weekTextLong: 'Vecka',
     allDayText: 'Heldag',
     moreLinkText: 'till',
     noEventsText: 'Inga händelser att visa',
+    closeHint: 'Stäng',
+    timeHint: 'Klockan',
+    eventHint: 'Händelse',
   };
 
   return sv;