--- a/InvenTree/company/models.py
+++ b/InvenTree/company/models.py
@@ -127,7 +127,10 @@ class Company(models.Model):
         upload_to=rename_company_image,
         null=True,
         blank=True,
-        variations={'thumbnail': (128, 128)},
+        variations={
+            'thumbnail': (128, 128),
+            'preview': (256, 256),
+        },
         delete_orphans=True,
         verbose_name=_('Image'),
     )