… InvenTree/InvenTree/static/bootstrap-table/bootstrap-table.min.css
… InvenTree/InvenTree/static/fullcalendar/locales-all.js
… InvenTree/InvenTree/static/fullcalendar/main.js
… InvenTree/InvenTree/static/fullcalendar/main.min.css
… InvenTree/company/migrations/0045_auto_20220616_1213.py
… InvenTree/company/templates/company/company_base.html
… InvenTree/company/templates/company/manufacturer_part.html
… InvenTree/company/templates/company/supplier_part.html
… InvenTree/part/migrations/0079_auto_20220616_1156.py
… InvenTree/part/serializers.py
… InvenTree/part/templates/part/detail.html
… InvenTree/part/templates/part/part_thumb.html
… InvenTree/stock/api.py
… InvenTree/stock/serializers.py
… InvenTree/stock/templates/stock/item_base.html
… InvenTree/templates/about.html
… InvenTree/templates/account/base.html
… InvenTree/templates/base.html
… InvenTree/templates/js/translated/build.js
… InvenTree/templates/page_base.html
… InvenTree/templates/skeleton.html
… InvenTree/templates/third_party_js.html
… InvenTree/users/models.py