- Closes https://github.com/inventree/InvenTree/issues/3173
- Closes https://github.com/inventree/InvenTree/issues/3168

### Summary of Changes

#### Refactoring "can_build" function

This function was extremely expensive in terms of number of database queries. It operated using a primitive for-loop structure, and has now been refactored to use a query annotation structure.

Using the demo dataset as an example, performance improvements are:

| Page | BOM Items | Query Count (Before) | Query Count (After) |
| --- | --- | --- | --- |
| /part/112 | 240 | **764** | 69 |
| /part/113 | 7 | 60 | 42 |
| /part/87 | 4 | 54 | 47 |
| /part/88 | 9 | 74 | 52 |
| /part/110 | 60 | 224 | 47 |
| /part/111 | 60 | 223 | 46 |

#### Minify CSS and JS Files

A number of static files were not sufficiently "minified"

##### Reduce size of part image

Instead of using the full-size part image, introduce a new "medium size" image preview, which can be sufficiently quicker to load

#### Lazy load of "About" dialog

Previously the "about" dialog was rendered on every page request - this is definitely *not* required! Removing this saved about ~400ms of server-side rendering time.
#### Caching Role Data

Store user role / permission data to memory cache

### Improve BOMItem API List Endpoint

- The BOMItem serializer provided the `user_detail` field which used the `PartBriefSerializer` class
- The `PartBriefSerializer` included a "stock" field, which ran an individual query for each part
- We already had stock information calculated using a subquery for each BomItem!

DB query improvements as below. API query has been reduced from O(n) to O(1)

| Bom Items | Before | After |
| --- | --- | --- |
| 3 | 25 | 22 |
| 4 | 26 | 22 |
| 7 | 30 | 18 |
| 9 | 35 | 25 |
| 60 | 83 | 23 | 
| 240 | 263 | 23 |

### Improve StockItem API List Endpoint

| Stock Items | Before | After |
| --- | --- | --- |
| 10 | 48 | 8 |
| 50 | 108 | 8 |
| 100 | 409 | 8 |
| 500 | 2013 | 9 |

<a href="https://gitpod.io/#https://github.com/inventree/InvenTree/pull/3185"><img src="https://gitpod.io/button/open-in-gitpod.svg"/></a>